requirements.txt
Django==3.2.3
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
requests==2.25.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
drf-yasg
pillow
django-cors-headers